포인트 사용/취소 개발

+++ src/main/java/com/plateer/ec1/promotion/mapper/PromotionTrxMapper.java
@@ -8,5 +8,5 @@ import org.apache.ibatis.annotations.Mapper;
 public interface PromotionTrxMapper {
     int insertDownloadCoupon(CouponRequestVo couponRequestVo);
 
-    int useCancelPoint(CcMbrPntModel ccMbrPntModel);
+    Long useCancelPoint(CcMbrPntModel ccMbrPntModel);
 }
